feat: choose submodule as a modal
